+++ CHANGELOG.rst
@@ -2,10 +2,17 @@
 Changelog
 =========
 
-Version 0.0.1
+Version 0.0.1 (2019-12-11)
 ===========
 
 - Bare minimum functionality
 - Executes notebook
 - Exports notebook to html
 - Attaches environment.yml and .ipynb to html
+
+
+Version 0.0.2 (2019-12-12)
+===========
+
+- Refactored convert module
+- Error handling for package list extraction